prueba 18 para el modulo de firma

+++ admin/Propiedades/Ver/DocumentosCanva/avisoPrivacidad.php
@@ -59,7 +59,11 @@ function avisoPrivacidad(
     }
     global $pdf;
     global $db;
-    unlink('../DocumentosFirmados/AvisoPrivacidad' . $idInmueble . '.pdf');
+
+    if (file_exists('../DocumentosFirmados/AvisoPrivacidad' . $idInmueble . '.pdf')){
+        unlink('../DocumentosFirmados/AvisoPrivacidad' . $idInmueble . '.pdf');
+    }
+
     $pdf = new PDF();
     $pdf->AddPage('P', 'Letter');
     $pdf->AddFont('Regular', '', 'Montserrat-Regular.php');
